+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Update ConsoleLogExporter.export to handle LogRecord's containing bytes type
+  in the body ([#4614](https://github.com/open-telemetry/opentelemetry-python/pull/4614/)).
 - opentelemetry-sdk: Fix invalid `type: ignore` that causes mypy to ignore the whole file
   ([#4618](https://github.com/open-telemetry/opentelemetry-python/pull/4618))
 - Add `span_exporter` property back to `BatchSpanProcessor` class